More aspect-oriented approach to IlsAwareTrait.
